requirements.txt
python-dotenv
requests
pydantic
openai
langchain
langgraph
tiktoken
transformers
sentence-transformers
faiss-cpu
beautifulsoup4
lxml
whisper
pyttsx3
sounddevice
SpeechRecognition
fastapi<0.104
uvicorn
streamlit
requests
pydantic<2.0.0
sentence-transformers
transformers
gtts
langchain

